refactor(score-calculator): the bars, walk and formula extract into a shared module
score-breakdown.js now owns the score/grade/plate arithmetic, both
attribution bars, the weighted-random next-letter walk and the shared
tooltip; phoenix-calculator.js imports it (the script tag becomes a
module - the import map already rewrites specifiers to hashed names)
and keeps everything page-specific: inputs, chips, toggles, the plays
dialog. The renderers take their strings from the caller, so each
surface localizes through its own channel. The calculator also learns
to prefill from query parameters - the Score Breakdown Dialog's
open-in-calculator link target (session-breakdown.md section 7.3).
