+++ ScoreTracker/ScoreTracker/App.razor
@@ -239,7 +239,10 @@
 <script src="@Assets["js/dashboard-grid.js"]"></script>
 <script src="@Assets["js/credential-storage.js"]"></script>
 <script src="@Assets["js/pumbility-calculator.js"]"></script>
-<script src="@Assets["js/phoenix-calculator.js"]"></script>
+@* A module, not a classic script: it imports the shared score-breakdown.js renderer that the
+   Score Breakdown Dialog also loads, so the two surfaces run one implementation. The import
+   map above rewrites the specifier to the content-hashed asset. *@
+<script type="module" src="@Assets["js/phoenix-calculator.js"]"></script>
 <script src="@Assets["js/challenge-board.js"]"></script>
 @* Ctrl+Alt+R cycles the reconnect overlay through its states; Ctrl+Alt+X clears it. The
    paused/resume-failed states can't be provoked on demand, so this is how they get seen. *@
