+++ ScoreTracker/ScoreTracker/wwwroot/js/phoenix-calculator.js
@@ -4,14 +4,15 @@
 // tooltips, and the plays dialog. Every constant this script reads — both mixes' grade floors,
 // the judgement weights, the owner-verified calorie table — comes from the JSON block the
 // server emitted from the engine, so nothing here can disagree with the page it sits under.
-// The score/plate/walk arithmetic mirrors ScoreScreen and is pinned there; without this script
-// the page loses the tool and keeps every fact.
-(function () {
-    'use strict';
+// The score/plate/walk/bars arithmetic lives in score-breakdown.js, shared with the Score
+// Breakdown Dialog so the two surfaces cannot drift (session-breakdown.md §7.3); without this
+// script the page loses the tool and keeps every fact.
+import { createScoreBreakdown, attachBreakdownTips } from './score-breakdown.js';
 
-    var page = document.querySelector('[data-sc-page]');
-    if (!page) return;
+var page = document.querySelector('[data-sc-page]');
+if (page) initialize();
 
+function initialize() {
     var constantsBlock = page.querySelector('[data-sc-constants]');
     if (!constantsBlock) return;
     var CONST = JSON.parse(constantsBlock.textContent);
@@ -19,39 +20,16 @@
     var OTHER_MIX = MIX === 'Phoenix2' ? 'Phoenix' : 'Phoenix2';
     var lang = document.documentElement.lang || undefined;
 
+    var engine = createScoreBreakdown(CONST, MIX, lang);
+    var score = engine.score;
+    var gradeFor = engine.gradeFor;
+    var plateFor = engine.plateFor;
+
     function n0(v) { return Math.round(v).toLocaleString(lang); }
     function fmt(template) {
         var args = Array.prototype.slice.call(arguments, 1);
         return template.replace(/\{(\d+)\}/g, function (_, i) { return args[+i] === undefined ? '' : args[+i]; });
     }
-
-    // ---- the formula, mirrored from ScoreScreen (floor; combo term at half a percent).
-    function score(p, gr, gd, bd, m, c) {
-        var total = p + gr + gd + bd + m;
-        if (total <= 0 || total >= 10000 || c > total || c < 0) return null;
-        var w = CONST.weights;
-        return Math.floor(
-            (w.accuracy * (p + w.great * gr + w.good * gd + w.bad * bd) + w.combo * c) / total * 1000000);
-    }
-
-    function gradeFor(value, mix) {
-        var floors = CONST.floors[mix];
-        for (var i = 0; i < floors.length; i++)
-            if (value >= floors[i].floor) return floors[i];
-        return floors[floors.length - 1];
-    }
-
-    function plateFor(gr, gd, bd, m) {
-        if (gr === 0 && gd === 0 && bd === 0 && m === 0) return 'Perfect Game';
-        if (gd === 0 && bd === 0 && m === 0) return 'Ultimate Game';
-        if (bd === 0 && m === 0) return 'Extreme Game';
-        if (m === 0) return 'Superb Game';
-        if (m <= 5) return 'Marvelous Game';
-        if (m <= 10) return 'Talented Game';
-        if (m <= 20) return 'Fair Game';
-        return 'Rough Game';
-    }
-
 
     // ---- the calculator.
     var calc = page.querySelector('[data-sc-calc]');
@@ -64,138 +42,48 @@
         return isNaN(v) ? 0 : Math.abs(v);
     }
 
-    // Deterministic per input: the walk reseeds so retyping the same screen retells the same
-    // path (mulberry32; production's stream-seeded Random differs per call, same idea).
-    function mulberry32(seed) {
-        var a = seed | 0;
-        return function () {
-            a = a + 0x6D2B79F5 | 0;
-            var t = Math.imul(a ^ a >>> 15, 1 | a);
-            t = t + Math.imul(t ^ t >>> 7, 61 | t) ^ t;
-            return ((t ^ t >>> 14) >>> 0) / 4294967296;
+    function judgementLabel(name) {
+        return inputs[name].previousElementSibling.textContent.trim();
+    }
+
+    // The shared renderers take their strings from the caller: this page's channel is the
+    // data-t-* attributes the server localized into the markup.
+    function barLabels(el) {
+        return {
+            perfects: judgementLabel('perfects'),
+            greats: judgementLabel('greats'),
+            goods: judgementLabel('goods'),
+            bads: judgementLabel('bads'),
+            misses: judgementLabel('misses'),
+            combo: judgementLabel('combo'),
+            brokenCombo: el.getAttribute('data-t-broken-combo'),
+            gained: el.getAttribute('data-t-gained'),
+            lost: el.getAttribute('data-t-lost'),
+            notEarned: el.getAttribute('data-t-not-earned'),
+            nothingLost: el.getAttribute('data-t-nothing-lost'),
+            clipped: el.getAttribute('data-t-clipped')
         };
     }
 
-    // The site's weighted-random walk (ScoreScreen.IterateWithWeightedRandom): one non-perfect
-    // converts per step, chosen in proportion to how many the player actually got, so the
-    // recipe auto-scales to their own error mix — a goods-heavy play is told "fewer goods".
-    function nextLetter(next, s) {
-        var current = gradeFor(s.score, MIX);
-        if (current.grade === 'SSS+') { next.textContent = next.getAttribute('data-t-best'); return; }
-        var floors = CONST.floors[MIX];
-        var target = null;
-        for (var i = floors.length - 1; i >= 0; i--) if (floors[i].floor > s.score) { target = floors[i]; break; }
-        if (!target) { next.textContent = next.getAttribute('data-t-best'); return; }
-
-        var rng = mulberry32(1949);
-        var w = { p: s.p, gr: s.gr, gd: s.gd, bd: s.bd, m: s.m, c: s.c };
-        var guard = 0;
-        while (guard++ < 200000) {
-            var sc = score(w.p, w.gr, w.gd, w.bd, w.m, w.c);
-            if (sc === null || gradeFor(sc, MIX).grade !== current.grade) break;
-            var total = w.m + w.bd + w.gd + w.gr;
-            if (total <= 0) { if (w.c >= w.p + w.gr) { guard = 1e9; break; } w.c++; continue; }
-            var pick = Math.floor(rng() * total) + 1;
-            if (pick > total - w.gr) { w.gr--; w.p++; }
-            else if (pick > total - w.gr - w.gd) { w.gd--; w.p++; w.c++; }
-            else if (pick > total - w.gr - w.gd - w.bd) { w.bd--; w.p++; w.c++; }
-            else { w.m--; w.p++; w.c++; }
-        }
-        var label = '<span class="sc-next-label">' + next.getAttribute('data-t-label') + ' · ' +
-            fmt(next.getAttribute('data-t-points'), n0(target.floor - s.score)) + '</span>';
-        if (guard >= 200000) { next.innerHTML = label; return; }
-        var got = gradeFor(score(w.p, w.gr, w.gd, w.bd, w.m, w.c), MIX).grade;
-        var parts = [];
-        if (w.p > s.p) parts.push(fmt(next.getAttribute('data-t-more-perfects'), '<b>' + (w.p - s.p) + '</b>'));
-        if (s.m > w.m) parts.push(fmt(next.getAttribute('data-t-fewer-misses'), '<b>' + (s.m - w.m) + '</b>'));
-        if (s.bd > w.bd) parts.push(fmt(next.getAttribute('data-t-fewer-bads'), '<b>' + (s.bd - w.bd) + '</b>'));
-        if (s.gd > w.gd) parts.push(fmt(next.getAttribute('data-t-fewer-goods'), '<b>' + (s.gd - w.gd) + '</b>'));
-        if (s.gr > w.gr) parts.push(fmt(next.getAttribute('data-t-fewer-greats'), '<b>' + (s.gr - w.gr) + '</b>'));
-        var comboGain = (s.m - w.m) + (s.bd - w.bd) + (s.gd - w.gd);
-        if (comboGain > 0) parts.push(fmt(next.getAttribute('data-t-more-combo'), '<b>' + comboGain + '</b>'));
-        next.innerHTML = label +
-            fmt(next.getAttribute('data-t-get'), '<b>' + got + '</b>') + ' ' + parts.join(', ') + '!';
+    function nextLabels(el) {
+        return {
+            label: el.getAttribute('data-t-label'),
+            points: el.getAttribute('data-t-points'),
+            best: el.getAttribute('data-t-best'),
+            get: el.getAttribute('data-t-get'),
+            morePerfects: el.getAttribute('data-t-more-perfects'),
+            fewerMisses: el.getAttribute('data-t-fewer-misses'),
+            fewerBads: el.getAttribute('data-t-fewer-bads'),
+            fewerGoods: el.getAttribute('data-t-fewer-goods'),
+            fewerGreats: el.getAttribute('data-t-fewer-greats'),
+            moreCombo: el.getAttribute('data-t-more-combo')
+        };
     }
 
-    function segment(cls, left, width, tip) {
-        return '<span class="sc-barseg ' + cls + '" data-sc-tip="' + tip.replace(/"/g, '&quot;') +
-            '" style="left:' + left + '%;width:' + width + '%"></span>';
-    }
-
-    function bars(el, s) {
-        var w = CONST.weights;
-        var total = s.p + s.gr + s.gd + s.bd + s.m;
-        var pts = function (count, weight) { return w.accuracy * weight * count / total * 1000000; };
-        // [class, label, points, count] — the bar is sized by POINTS, and a bad pays half a
-        // good per note, so the count rides every tooltip to keep the sliver legible.
-        var gained = [
-            ['judg-perfect', inputs.perfects.previousElementSibling.textContent.trim(), pts(s.p, 1), s.p],
-            ['judg-great', inputs.greats.previousElementSibling.textContent.trim(), pts(s.gr, w.great), s.gr],
-            ['judg-good', inputs.goods.previousElementSibling.textContent.trim(), pts(s.gd, w.good), s.gd],
-            ['judg-bad', inputs.bads.previousElementSibling.textContent.trim(), pts(s.bd, w.bad), s.bd],
-            ['sc-seg-combo', inputs.combo.previousElementSibling.textContent.trim(), w.combo * s.c / total * 1000000, s.c]
-        ];
-        var lost = [
-            ['judg-great', inputs.greats.previousElementSibling.textContent.trim(), pts(s.gr, 1 - w.great), s.gr],
-            ['judg-good', inputs.goods.previousElementSibling.textContent.trim(), pts(s.gd, 1 - w.good), s.gd],
-            ['judg-bad', inputs.bads.previousElementSibling.textContent.trim(), pts(s.bd, 1 - w.bad), s.bd],
-            ['judg-miss', inputs.misses.previousElementSibling.textContent.trim(), pts(s.m, 1), s.m],
-            ['sc-seg-combo', el.getAttribute('data-t-broken-combo'), w.combo * (total - s.c) / total * 1000000, total - s.c]
-        ];
-        var lo = Math.min(Math.floor(s.score / 100000), 9) * 100000;
-        var span = 1000000 - lo;
-        var acc = 0;
-        var segs1 = '';
-        gained.forEach(function (g) {
-            var a = Math.max(acc, lo);
-            var b = Math.min(acc + g[2], 1000000);
-            if (b > a) {
-                var clipped = acc < lo;
-                segs1 += segment(g[0] + (clipped ? ' sc-clip' : ''), (a - lo) / span * 100, (b - a) / span * 100,
-                    g[1] + ' ×' + n0(g[3]) + ' — +' + n0(g[2]) +
-                    (clipped ? ' (' + el.getAttribute('data-t-clipped') + ')' : ''));
-            }
-            acc += g[2];
+    function markSpreadRow(gradeName) {
+        page.querySelectorAll('[data-sc-spread-grade]').forEach(function (row) {
+            row.classList.toggle('sc-you', row.getAttribute('data-sc-spread-grade') === gradeName);
         });
-        var missing = 1000000 - s.score;
-        var leg1 = gained.filter(function (g) { return g[2] >= .5; }).map(function (g) {
-            return '<span class="sc-it"><i class="sc-sw sc-barseg ' + g[0] +
-                '" style="position:static;width:9px;height:9px;border-radius:50%"></i>' +
-                g[1] + ' <b>+' + n0(g[2]) + '</b></span>';
-        }).join('');
-        if (missing > 0)
-            leg1 += '<span class="sc-it"><i class="sc-sw sc-sw-empty"></i>' +
-                el.getAttribute('data-t-not-earned') + ' <b>' + n0(missing) + '</b></span>';
-
-        var totalLost = 0;
-        lost.forEach(function (l) { totalLost += l[2]; });
-        var segs2 = '';
-        var acc2 = 0;
-        if (totalLost > 0)
-            lost.forEach(function (l) {
-                if (l[2] <= 0) return;
-                segs2 += segment(l[0], acc2 / totalLost * 100, Math.max(l[2] / totalLost * 100 - .3, .3),
-                    l[1] + ' ×' + n0(l[3]) + ' — −' + n0(l[2]) + ' / ' + n0(totalLost) +
-                    ' (' + Math.round(l[2] / totalLost * 100) + '%)');
-                acc2 += l[2];
-            });
-        var leg2 = totalLost > 0
-            ? lost.filter(function (l) { return l[2] >= .5; }).map(function (l) {
-                return '<span class="sc-it"><i class="sc-sw sc-barseg ' + l[0] +
-                    '" style="position:static;width:9px;height:9px;border-radius:50%"></i>' +
-                    l[1] + ' <b>−' + n0(l[2]) + '</b></span>';
-            }).join('')
-            : '<span class="sc-it">' + el.getAttribute('data-t-nothing-lost') + '</span>';
-
-        el.innerHTML =
-            '<div class="sc-bar"><h4>' + el.getAttribute('data-t-gained') +
-            '<span class="sc-range">' + n0(lo) + ' → 1,000,000</span></h4>' +
-            '<div class="sc-bartrack">' + segs1 + '</div>' +
-            '<div class="sc-barlegend">' + leg1 + '</div></div>' +
-            '<div class="sc-bar"><h4>' + el.getAttribute('data-t-lost') +
-            '<span class="sc-range">0 → ' + n0(totalLost) + '</span></h4>' +
-            '<div class="sc-bartrack">' + segs2 + '</div>' +
-            '<div class="sc-barlegend">' + leg2 + '</div></div>';
     }
 
     function calorieSteps(calories, total) {
@@ -207,12 +95,6 @@
             perStep = .035 + .0023 * bucket;
         }
         return calories / perStep;
-    }
-
-    function markSpreadRow(gradeName) {
-        page.querySelectorAll('[data-sc-spread-grade]').forEach(function (row) {
-            row.classList.toggle('sc-you', row.getAttribute('data-sc-spread-grade') === gradeName);
-        });
     }
 
     function recalc() {
@@ -255,9 +137,9 @@
             ? fmt(cross.getAttribute('data-t-other'), cross.getAttribute('data-other-mix'), other.grade)
             : '';
         nextEl.hidden = false;
-        nextLetter(nextEl, s);
+        engine.renderNextLetter(nextEl, s, nextLabels(nextEl));
         barsEl.hidden = false;
-        bars(barsEl, s);
+        engine.renderBars(barsEl, s, barLabels(barsEl));
         var calories = parseFloat(inputs.calories.value);
         if (!isNaN(calories) && calories > 0) {
             calEl.innerHTML = fmt(calEl.getAttribute('data-t-arrows'),
@@ -271,10 +153,21 @@
 
     Object.keys(inputs).forEach(function (name) { inputs[name].addEventListener('input', recalc); });
 
+    // A link can hand the calculator a play: the Score Breakdown Dialog's "open in the
+    // calculator" carries the counts as query parameters, and typing over them works exactly
+    // as if they had been typed in the first place.
+    var params = new URLSearchParams(window.location.search);
+    var prefilled = false;
+    ['perfects', 'greats', 'goods', 'bads', 'misses', 'combo'].forEach(function (name) {
+        var v = params.get(name);
+        if (v !== null && /^\d+$/.test(v)) { inputs[name].value = v; prefilled = true; }
+    });
+    if (prefilled) recalc();
+
     // ---- the chart-size chips reprice the cost cards and the budget table.
     var sizeGroup = page.querySelector('[data-sc-sizes]');
     if (sizeGroup) {
-        function costOf(kind, notes) {
+        var costOf = function (kind, notes) {
             // Best-case combos, mirroring the model: greats keep the run, a good tops out one
             // short, a bad or an edge miss breaks one note off it.
             if (kind === 'great') return 1000000 - score(notes - 1, 1, 0, 0, 0, notes);
@@ -282,8 +175,8 @@
             if (kind === 'bad') return 1000000 - score(notes - 1, 0, 0, 1, 0, notes - 1);
             if (kind === 'miss') return 1000000 - score(notes - 1, 0, 0, 0, 1, notes - 1);
             return 1000000 - score(notes - 1, 0, 0, 0, 1, Math.floor((notes - 1) / 2));
-        }
-        function budgetFor(floor, notes) {
+        };
+        var budgetFor = function (floor, notes) {
             var low = 0, high = notes;
             while (low < high) {
                 var candidate = Math.floor((low + high + 1) / 2);
@@ -291,7 +184,7 @@
                 else high = candidate - 1;
             }
             return low;
-        }
+        };
         sizeGroup.addEventListener('click', function (e) {
             var chip = e.target.closest('[data-sc-size]');
             if (!chip) return;
@@ -340,16 +233,16 @@
         var playFilter = dialog.querySelector('[data-sc-play-filter]');
         var plays = null;
 
-        function esc(text) {
+        var esc = function (text) {
             return String(text).replace(/&/g, '&amp;').replace(/</g, '&lt;').replace(/>/g, '&gt;')
                 .replace(/"/g, '&quot;');
-        }
+        };
 
-        function messageRow(text) {
+        var messageRow = function (text) {
             return '<tr><td style="color:var(--mix-ink-muted)">' + esc(text) + '</td></tr>';
-        }
+        };
 
-        function renderPlays() {
+        var renderPlays = function () {
             if (!plays) { playList.innerHTML = messageRow(playList.getAttribute('data-t-loading')); return; }
             if (plays.length === 0) { playList.innerHTML = messageRow(playList.getAttribute('data-t-empty')); return; }
             var query = playFilter.value.trim().toLowerCase();
@@ -385,9 +278,9 @@
                     '</span></span></td></tr>';
             });
             playList.innerHTML = rows.length ? rows.join('') : messageRow(playList.getAttribute('data-t-none-match'));
-        }
+        };
 
-        function openDialog() {
+        var openDialog = function () {
             dialog.hidden = false;
             playFilter.value = '';
             renderPlays();
@@ -400,9 +293,9 @@
                     renderPlays();
                 })
                 .catch(function () { plays = []; renderPlays(); });
-        }
+        };
 
-        function closeDialog() { dialog.hidden = true; }
+        var closeDialog = function () { dialog.hidden = true; };
 
         loadButton.addEventListener('click', openDialog);
         dialog.querySelector('[data-sc-dialog-close]').addEventListener('click', closeDialog);
@@ -425,18 +318,7 @@
         });
     }
 
-    // ---- one tooltip for every [data-sc-tip] carrier (chart hit areas, bar segments).
-    var tip = document.createElement('div');
-    tip.className = 'sc-tip';
-    tip.hidden = true;
-    document.body.appendChild(tip);
-    page.addEventListener('mousemove', function (e) {
-        var carrier = e.target.closest('[data-sc-tip]');
-        if (!carrier) { tip.hidden = true; return; }
-        tip.textContent = carrier.getAttribute('data-sc-tip');
-        tip.hidden = false;
-        tip.style.left = Math.min(e.clientX + 14, window.innerWidth - 310) + 'px';
-        tip.style.top = Math.min(e.clientY + 14, window.innerHeight - 90) + 'px';
-    });
-    page.addEventListener('mouseleave', function () { tip.hidden = true; });
-})();
+    // ---- one tooltip for every [data-sc-tip] carrier (chart hit areas, bar segments) —
+    // shared with the breakdown dialog so a page hosting both never grows two.
+    attachBreakdownTips(page);
+}
